Read Hobo CSVs

Read Hobo CSVs

read_hobo_csv(file = ".", orders = c("Ymd HMS", "dmy HMS"),
  units = "degC", tz = "Etc/GMT+8", recursive = FALSE, quiet = FALSE)

Arguments

file
A string of the file or directory.

orders
A character vector of date-time formats used by parse_date_time.

units
A string of the units to convert the temperature data to using ud.convert.

tz
A string of the time zone to convert the date times to using with_tz.

recursive
A flag indicating whether to read files from subdirectories.

quiet
A flag indicating whether to provide messages. Ignored if file is a file (as opposed to a directory).

Value

A tibble of the data with the temperature.

Examples

read_hobo_csv(system.file("hobo", "10723440.csv", package = "poisutils"))

#> imported 11 rows of data from '/Users/joe/Dropbox (Poisson Consulting)/Code/poisutils/poisutils/inst/hobo/10723440.csv'

#> # A tibble: 11 x 6
#>      Logger            DateTime Temperature_degC FileRow FileName
#>       <chr>              <dttm>            <dbl>   <int>    <chr>
#>  1 10723440 2015-11-10 07:18:12           14.649       1 10723440
#>  2 10723440 2015-11-10 07:23:12           17.106       3 10723440
#>  3 10723440 2015-11-10 07:28:12           16.177       4 10723440
#>  4 10723440 2015-11-10 07:33:12           13.112       5 10723440
#>  5 10723440 2015-11-10 07:38:12           10.883       6 10723440
#>  6 10723440 2015-11-10 07:43:12            9.435       7 10723440
#>  7 10723440 2015-11-10 07:48:12            8.419       8 10723440
#>  8 10723440 2015-11-10 07:53:12            7.795       9 10723440
#>  9 10723440 2015-11-10 07:58:12            7.393      10 10723440
#> 10 10723440 2016-04-09 04:38:12            3.932      11 10723440
#> 11 10723440 2016-04-09 04:43:12            3.932      12 10723440
#> # ... with 1 more variables: Directory <chr>